test(backends): autoload data once per session for pytest backend tests
If the database already exists, the database will be dropped, and
recreated.

Expected that every backend will have a `load_data` static function
in the `TestConf` in ibis/backends/{backend}/tests/conftest.py.

Moved relevant helper functions to ibis/backends/conftest.py.

As of this code, `ci/datamgr.py load {backend}` will still work.

Used filelock.FileLock to ensure databases instantiate only once per pytest invocation
Handling for multiple pytest threads run concurrently.
Avoiding a situation where each thread idependently tries to instantiate
the database.

Closes #2952
